+++ ui/src/app/pages/appAnalysis/app-selector.tsx
@@ -1,18 +1,11 @@
 import * as React from 'react';
 import { useState } from 'react';
-import { Dropdown } from 'primereact/dropdown';
 
 import { BillingStatus } from 'generated/fetch';
 
 import { UIAppType } from 'app/components/apps-panel/utils';
 import { Button } from 'app/components/buttons';
 import { rstudioConfigIconId } from 'app/components/help-sidebar-icons';
-import {
-  Modal,
-  ModalBody,
-  ModalFooter,
-  ModalTitle,
-} from 'app/components/modals';
 import { NewJupyterNotebookModal } from 'app/pages/analysis/new-jupyter-notebook-modal';
 import colors from 'app/styles/colors';
 import { reactStyles } from 'app/utils';
@@ -21,22 +14,15 @@ import { setSidebarActiveIconStore } from 'app/utils/navigation';
 import { WorkspaceData } from 'app/utils/workspace-data';
 import { WorkspacePermissionsUtil } from 'app/utils/workspace-permissions';
 
+import { AppSelectorModal } from './app-selector-modal';
+
 const styles = reactStyles({
   startButton: {
     paddingLeft: '0.75rem',
     height: '3rem',
     backgroundColor: colors.secondary,
   },
-  appsLabel: {
-    color: colors.primary,
-    fontWeight: 600,
-    fontSize: '14px',
-    lineHeight: '24px',
-    paddingBottom: '0.75rem',
-  },
 });
-
-export const APP_LIST = [UIAppType.JUPYTER, UIAppType.RSTUDIO];
 
 const enum VisibleModal {
   None = 'None',
@@ -92,47 +78,9 @@ export const AppSelector = (props: AppSelectorProps) => {
         <div style={{ width: '9rem', paddingLeft: '1rem' }}>Choose an App</div>
       </Button>
       {visibleModal === VisibleModal.SelectAnApp && (
-        <Modal
-          data-test-id='select-application-modal'
-          aria={{
-            label: 'Select Applications Modal',
-          }}
-        >
-          <ModalTitle>Analyze Data</ModalTitle>
-          <ModalBody>
-            <div style={styles.appsLabel}>Select an application</div>
-            <Dropdown
-              id='application-list-dropdown'
-              data-test-id='application-list-dropdown'
-              ariaLabel='Application List Dropdown'
-              value={selectedApp}
-              appendTo='self'
-              options={APP_LIST}
-              placeholder='Choose One'
-              onChange={(e) => setSelectedApp(e.value)}
-              style={{ width: '13.5rem' }}
-            />
-          </ModalBody>
-          <ModalFooter style={{ paddingTop: '3rem' }}>
-            <Button
-              style={{ marginRight: '3rem' }}
-              type='secondary'
-              aria-label='close'
-              onClick={onClose}
-            >
-              Close
-            </Button>
-            <Button
-              data-test-id='next-btn'
-              type='primary'
-              aria-label='next'
-              onClick={onNext}
-              disabled={!selectedApp}
-            >
-              Next
-            </Button>
-          </ModalFooter>
-        </Modal>
+        <AppSelectorModal
+          {...{ selectedApp, setSelectedApp, onNext, onClose }}
+        />
       )}
       {visibleModal === VisibleModal.Jupyter && (
         <NewJupyterNotebookModal
